+++ CHANGELOG.md
@@ -20,6 +20,8 @@ On this page we provide a summary of the main API changes, new features and exam
 - Add `build_objects` parameter to `CeedOperatorLinearAssembleQFunctionBuildOrUpdateFallback` to allow for passing uninitialized vectors and restrictions
 - Move JiT helper functions only required by SYCL backends to `ceed/jit-tools-deprecated.h`.
 These functions will be removed when the SYCL backends are updated to reflect the improvements in the CUDA and HIP backends.
+- Add `CeedOperatorLinearAssembleGetFlopsEstimate`, `CeedOperatorLinearAssembleDiagonalGetFlopsEstimate`, and
+  `CeedOperatorLinearAssemblePointBlockDiagonalGetFlopsEstimate` to estimate FLOPs for linear operator assembly.
 
 ### New features